+++ src/Adjustments/Changelog.md
@@ -10,7 +10,7 @@
 - Dropped Enum v3 Support
 - Added PHP 8.3 Support
 - Changed minimum Laravel version to v10.38.2
-- Changed minimal Enum requirement to v4.1
+- Changed minimal Enum requirement to v4.2
 - Added the `mapInto()` method to the `RelationAdjustmentCollection` class, which forwards the call to the underlying Eloquent collection
 - BC: Added the `deleteByType()` method to the `AdjustmentCollection` interface + both implementation